backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.2
transformers==4.35.2
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
scikit-learn==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
nltk==3.8.1
textblob==0.17.1
